Converted to 64-bit

+++ src/include/string.h
@@ -11,7 +11,7 @@ void *memset32(void *dest, u32 value, u32 count);
 int memcmp(const void *dest, const void *src, u32 count);
 void *strcpy(char *dest, const char *src);
 int strcmp(const char *dest, const char *src);
-unsigned long strlen(const char *s);
+u32 strlen(const char *s);
 char *strcat(char *dest, const char *src);
 char *strchr(const char *str, char c);
 char *strrchr(const char *str, char c);
